Use correct integer type for complexity. MSVC warns otherwise

+++ include/benchmark/benchmark_api.h
@@ -367,7 +367,7 @@ public:
   }
 
   BENCHMARK_ALWAYS_INLINE
-  size_t complexity_length_n() {
+  int complexity_length_n() {
     return complexity_n_;
   }
